Extract `assertRequestAuthentication` into own file
