+++ apps/web/app/api/(oauth)/token/token.test.ts
@@ -1,12 +1,11 @@
 /** @jest-environment node */
 import { OAuth2ErrorCode } from '@/lib/oauth/error';
-import { assertPKCECodeChallenge, assertRequestAuthentication, handleTokenRequest } from './token';
+import { assertPKCECodeChallenge, handleTokenRequest } from './token';
 import { dbMock } from '@/lib/db.mock';
-import { Account, Authorization, Client, ClientSecret, ClientType } from '@gw2me/database';
+import { Account, Authorization, Client } from '@gw2me/database';
 import { expiresAt } from '@/lib/date';
 import { Scope } from '@gw2me/client';
 import { describe, expect, it } from '@jest/globals';
-import { generateClientSecretAndHash } from '@/lib/token';
 
 type MockAuthorization = Authorization & { accounts: Pick<Account, 'id'>[] };
 
@@ -27,14 +26,13 @@ const mockAuthorization: MockAuthorization = {
   accounts: [],
 };
 
-const client: Client & { secrets: ClientSecret[] } = {
+const client: Client = {
   id: 'test',
   applicationId: 'app-id',
   callbackUrls: [],
   createdAt: new Date(),
   updatedAt: new Date(),
   type: 'Public',
-  secrets: [],
 };
 
 const headers = new Headers();
@@ -103,29 +101,5 @@ describe('/api/token', () => {
     it('unknown method', () => expect(() => assertPKCECodeChallenge('foo:challenge', 'bar')).toThrowOAuth2Error(OAuth2ErrorCode.invalid_request, 'Unsupported code challenge'));
     it('wrong verifier', () => expect(() => assertPKCECodeChallenge('S256:challenge', 'wrong')).toThrowOAuth2Error(OAuth2ErrorCode.invalid_request, 'code challenge verification failed'));
     it('success', () => expect(() => assertPKCECodeChallenge('S256:w6uP8Tcg6K2QR905Rms8iXTlksL6OD1KOWBxTK7wxPI', 'foobar')).not.toThrow());
-  });
-
-  describe('authentication', () => {
-    describe('public client', () => {
-      it('works', () => expect(() => assertRequestAuthentication(client, headers, {})).not.toThrow());
-      it('throws if authorization header was provided', () => expect(() => assertRequestAuthentication(client, new Headers({ 'Authorization': 'Basic foo' }), {})).toThrowOAuth2Error(OAuth2ErrorCode.invalid_request, 'Do not pass authorization for public clients.'));
-      it('throws if client secret was provided', () => expect(() => assertRequestAuthentication(client, headers, { client_secret: '' })).toThrowOAuth2Error(OAuth2ErrorCode.invalid_request, 'Do not pass authorization for public clients.'));
-    });
-
-    describe('confidential client', () => {
-      let secret: Awaited<ReturnType<typeof generateClientSecretAndHash>>;
-      let confidentialClient: typeof client;
-
-      beforeAll(async () => {
-        secret = await generateClientSecretAndHash();
-        confidentialClient = { ...client, type: ClientType.Confidential, secrets: [{ secret: secret.hashed } as ClientSecret] };
-      });
-
-      it('throws if no authentication was provided', () => expect(() => assertRequestAuthentication(confidentialClient, headers, {})).toThrowOAuth2Error(OAuth2ErrorCode.invalid_request, 'Missing authorization for confidential client'));
-      it('handles non basic auth header', () => expect(() => assertRequestAuthentication(confidentialClient, new Headers({ 'Authorization': 'Bearer foo' }), {})).toThrowOAuth2Error(OAuth2ErrorCode.invalid_request, 'Only "Basic" authorization is supported using the Authorization header'));
-      it('handles malformed basic auth header', () => expect(() => assertRequestAuthentication(confidentialClient, new Headers({ 'Authorization': 'Basic !"§$' }), {})).toThrowOAuth2Error(OAuth2ErrorCode.invalid_request, 'Unexpected client_id in Authorization header'));
-      it('supports client_secret_basic', () => expect(() => assertRequestAuthentication(confidentialClient, new Headers({ 'Authorization': `Basic ${btoa(`${client.id}:${secret.raw}`)}` }), {})).not.toThrow());
-      it('supports client_secret_post', () => expect(() => assertRequestAuthentication(confidentialClient, headers, { client_secret: secret.raw })).not.toThrow());
-    });
   });
 });
